Simplified current code
This simplifies the code for all of the abstract types. This push will also allow for the storing of images within the software allowing for some way to display the images you are looking for. All that is left to have the software display them in some way on the screen. I have testing I was able to save one of the screen shots and it looked good On my end